Added multiple data sets
The datasets ending with _main are subnetworks derived from facebook_ucsd.edge, where all their nodes are accessible to each other.
It turns out that the main graph contains a huge subnet that is ~99% the total size, and other tiny (2-4 node) subnets. Thus, analyzing the main subnet is probably accurate enough to measure the connectedness of its communities.